module2_prompt_engineering/requirements.txt
litellm
python-dotenv